+++ source/lib64/ext-lib-bind-src/v8_d/util8json.h
@@ -39,6 +39,6 @@ void double_to_mantissa_exponent(double inp, int64_t *mantissa, int64_t *exponen
 string exponent_and_mantissa_to_string(long decimal_mantissa_data, long decimal_exponent_data);
 bool jsobject_log(Local<Value> value);
 template < typename T > std::string to_string( const T& n );   
-int to_int(char const *s);
+int64_t to_int(char const *s);
 
 #endif
